adding migrations

+++ Migrations/AppDbContextModelSnapshot.cs
@@ -29,8 +29,8 @@ namespace RepositoryPattern.Migrations
 
                     SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
-                    b.Property<decimal>("Price")
-                        .HasColumnType("DECIMAL");
+                    b.Property<double>("Price")
+                        .HasColumnType("float");
 
                     b.Property<string>("Title")
                         .IsRequired()
